Refactor auth methods into testutils
