+++ src/server/auth/server/testing/oidc_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/pachyderm/pachyderm/src/client"
 	"github.com/pachyderm/pachyderm/src/client/auth"
 	"github.com/pachyderm/pachyderm/src/client/pkg/require"
+	tu "github.com/pachyderm/pachyderm/src/server/pkg/testutil"
 )
 
 // Dex's mock connector always returns the same identity for all authentication
@@ -43,8 +44,8 @@ func TestOIDCAuthCodeFlow(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skipping integration tests in short mode")
 	}
-	deleteAll(t)
-	adminClient, testClient := getPachClient(t, admin), getPachClient(t, "")
+	tu.DeleteAll(t)
+	adminClient, testClient := tu.GetAuthenticatedPachClient(t, tu.AdminUser), tu.GetAuthenticatedPachClient(t, "")
 
 	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
 		&auth.SetConfigurationRequest{Configuration: OIDCAuthConfig})
@@ -96,7 +97,7 @@ func TestOIDCAuthCodeFlow(t *testing.T) {
 	require.Equal(t, "idp:"+dexMockConnectorEmail, whoAmIResp.Username)
 	require.False(t, whoAmIResp.IsAdmin)
 
-	deleteAll(t)
+	tu.DeleteAll(t)
 }
 
 // TestOIDCTrustedApp tests using an ID token issued to another OIDC app to authenticate.
@@ -107,8 +108,8 @@ func TestOIDCTrustedApp(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skipping integration tests in short mode")
 	}
-	deleteAll(t)
-	adminClient, testClient := getPachClient(t, admin), getPachClient(t, "")
+	tu.DeleteAll(t)
+	adminClient, testClient := tu.GetAuthenticatedPachClient(t, tu.AdminUser), tu.GetAuthenticatedPachClient(t, "")
 
 	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
 		&auth.SetConfigurationRequest{Configuration: OIDCAuthConfig})
@@ -143,7 +144,7 @@ func TestOIDCTrustedApp(t *testing.T) {
 	// idp:admin is an admin of the IDP but not Pachyderm
 	require.False(t, whoAmIResp.IsAdmin)
 
-	deleteAll(t)
+	tu.DeleteAll(t)
 }
 
 // Rewrite the Location header to point to the returned path at `host`
